BackEnd/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
GitPython==3.1.30
requests==2.22.0
